refactor: decompose large functions into focused modules
Split SQLite migrate into ordered step functions; extract scheduler helpers
for market cache, fetch groups, pruning, and listing pipeline; replace bot
callback switch with a registry; factor createSearch and health snapshot;
partition locale strings by domain.

Closes #433, #434, #435, #436, #437, #438, #439, #440, #442, #445, #446, #447, #451
